Feature Additions
1) Fixed the code for starting the User Profile Synchronization Service.
Let me know if you run into any issues with this not starting
automatically again.

2) Fixed the code that starts Access Database Services 2010 service in
2013 builds. This was not starting due to the difference in the service
name between 2010 and 2013. It should be starting properly now.

3) Added some checking code for the new SharePoint Tools folder copy
that I added yesterday. It will no longer throw an error if the folder
already exists which can result from multiple run throughs of the auto
build script on the same server.

4) Added a prompt in the data collection phase of the script
(SetVars.ps1) for installing 2010 or 2013. This will allow me to add
2013 services to the appropriate servers in the FarmConfig.xml file.

5) Added code to Create the App Management and Subscription Settings
Service Apps for 2013 builds including the ability to isolate these
Service Apps into their own App Pools if requested by the customer.

6) Added two new App Pool Accounts to allow for isolation of the above
two Service Apps (SP_AppMgmt_AP and PROD_SP_Subscr_AP). If High
Isolation is not chosen during the data collection phase then these two
apps will use the default App Pool account as usual.

7) Added code to start App Management Service and Subscription Setting
service for 2013 builds.

8) Added code to request the App Domain and App Prefix during the data
collection stage and automatically set those settings during the
post-install config phase. If this data is not provided then they just
won't be configured and will have to be done by hand sometime after the
build either through script or through CA. Not providing this
information will not affect the provisioning of the App Mgmt and
Subscription Settings Service Apps.

9) Re-fixed the custom SQL Alias creation code block. I swear I fixed
this once but it was before I started tracking with github and it
appears to have regressed somewhere along the way. If you choose a
non-default SQL alias during the data collection phase it will correctly
create the alias in the registry and use it when building the farm.