Bugfixes
1) RS_AutoBuildSetServAccts.ps1
a. Added OS detection and “Import-Module ActiveDirectory” for
pre-Win2012.
i. This loads the AD CMDLets once the Remote Admin Tools Role is
installed
ii. Win 2012 loads this module automatically
b. Added a Try-Catch around connecting to AD with Get-ADRootDSE to
ensure good feedback if the connection fails
i. In the catch block I added some recommendations for troubleshooting
if the connection fails
2) RS_AutoBuildPost.ps1
a. Added SharePoint SKU detection to the AppFabric CU update code so it
only executes for 2013 installs.
b. Added an additional nested Try/Catch block around the code for
disabling the health analyser rules so it tests for the current status
of the rule before attempting to disable.
c.  Added code for Checking/Setting Search Service Performance Level
i. The way I set this up is to do a compare between the list of
Distributed Cache servers in the XML and the list of Index Component
Servers in the XML and only reduce the Performance Level if the
Distributed Cache list contains the same servers from the Index
Component list
1. It’s not a perfect solution, but the logic here is that the Index
Component will always be in the list of Distributed Cache servers unless
the Index Component is split off to its own server
a. If the Index Component servers are not in the Distributed Cache
servers list, Search will remain at Maximum
3) RS_AutoBuild.ps1
a. Corrected a typo in the SQL Connection test code which only surfaced
when specifying a custom port in the XML
4) AutoSPInstallerMain.ps1
a. Corrected the function call for prepping the XML for Foundation
installs with AutoSPInstaller
5) AutoSPInstallerFunctionsCustom.ps1
a. Removed the need for the AutSPInstallerInput.xml file path to be
passed to the PrepFoundation function and use the available environment
variables instead.
6) Upgrading AutoSPInstaller.txt
a. Updated the AutoSPInstaller upgrade instructions with the correct
code that needs to be re-added after upgrade to allow for Foundation
installs.